+++ lib/v7cms/tasks.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+# Load v7cms rake tasks
+load File.expand_path('tasks/v7cms.rake', __dir__)